fix(ipsec): 第二批协议正确性修复——msgid 32 位、tick 断链、在途请求互斥、INFORMATIONAL 响应验证
- msgid/pending_msgid 全链路 8 位改 32 位 (RFC 7296 单调不回绕),
  修复 DPD 每 30s 消耗一个 ID 导致约 2.1 小时回绕被 strongSwan
  按 replay 丢弃而拆隧道的问题
- tick 定时器 transport_err 分支不再无条件 return: MOBIKE socket
  重开成功路径恢复 re-arm, DPD/rekey/地址轮询不再永久停摆
- 新增在途请求互斥 (ipsec_ike_busy): DPD/MOBIKE 更新/CREATE_CHILD_SA
  任一在途时其余让路, 下一秒 tick 自动重试; ipsec_start_retrans
  先摘旧定时器再 arm, 杜绝并存重传定时器
- INFORMATIONAL 响应 (DPD/MOBIKE ack) 经 SK 解密 + ICV 校验后才
  采信, 未认证响应不再能伪造保活; 认证通过后恢复响应路径的
  MOBIKE 地址学习 (RFC 4555 §3)
- docs/ipsec-design.md 同步单在途槽位与响应验证说明

验证: PC 模拟器编译通过; netdrv_ipsec_basic 对真实网关
3 passed / 0 failed; 75s 空闲挂测两轮 DPD 正常采信

+++ components/network/ipsec/include/ipsec/ipsec_ike.h
@@ -215,8 +215,8 @@ struct ipsec_client {
 
     /* IKE state */
     uint8_t  phase;
-    uint8_t  msgid;               /* next request message id */
-    uint8_t  pending_msgid;
+    uint32_t msgid;               /* next request message id (RFC 7296: 32-bit, monotonic) */
+    uint32_t pending_msgid;
     uint8_t  last_exchange;       /* exchange type of the pending request */
     uint8_t  nat_detected;
     uint8_t  cookie_len;
